refactor(LottoPrize) : LottoStatistics 클래스에서 Lotto 상금과 관련된 클래스 분리
