+++ src/main/java/lotto/controller/LottoController.java
@@ -12,6 +12,7 @@ public class LottoController {
     private static final OutputView outputView = new OutputView();
     private List<Lotto> lottoTicketList = new ArrayList<>();
     private LottoStatistics lottoStatistics = new LottoStatistics();
+    private LottoPrize lottoPrize = new LottoPrize();
 
     public static LottoController create(){
         return new LottoController();
@@ -22,7 +23,7 @@ public class LottoController {
         issueLotto(lottoPrice.lottoTicketCount());
         lottoStatistics.calculation(lottoTicketList,inputLottoNumbers());
         printStatistics(lottoStatistics.getLottoStatisticsList());
-        printEarning(lottoStatistics.calculateEarning(lottoPrice.getLottoPrice()));
+        printEarning(lottoPrize.calculateEarning(lottoPrice.getLottoPrice()));
     }
     private LottoPrice buyLottoTicket(){
         return new LottoPrice(inputView.inputPrice());
